+++ scrapli/driver/core/__init__.py
@@ -3,6 +3,7 @@ from scrapli.driver.core.arista_eos.driver import EOSDriver
 from scrapli.driver.core.cisco_iosxe.driver import IOSXEDriver
 from scrapli.driver.core.cisco_iosxr.driver import IOSXRDriver
 from scrapli.driver.core.cisco_nxos.driver import NXOSDriver
+from scrapli.driver.core.generic_drivery import GenericDriver
 from scrapli.driver.core.juniper_junos.driver import JunosDriver
 
 __all__ = (
@@ -10,5 +11,6 @@ __all__ = (
     "IOSXEDriver",
     "IOSXRDriver",
     "NXOSDriver",
+    "GenericDriver",
     "JunosDriver",
 )