+++ CHANGELOG.md
@@ -7,3 +7,8 @@ CHANGELOG
 - Add an `atexit` function for the ssh2 transport which forcibly closes the connection. This fixes a bug where if a
  user did not manually close the connection (or use a context manager for the connection) the script would hang open
   until an interrupt.
+- Added a `GenericDriver` for those with non-core platforms. The `GenericDriver` has a really broad prompt pattern
+ match, doesn't know about privilege levels or any other device specific stuff, but does provide the `send_command
+ `, `send_commands`, `send_interact`, and `get_prompt` methods just like the "core" drivers do. This should be a
+  decent starting point for anyone working on non-core platforms!
+- Minor unit test improvement to cover send_commands (plural) and to cover the new `GenericDriver`